+++ src/main/java/insane96mcp/insanelib/network/NetworkHandler.java
@@ -1,21 +1,26 @@
 package insane96mcp.insanelib.network;
 
 import insane96mcp.insanelib.InsaneLib;
-import net.minecraft.resources.ResourceLocation;
+import insane96mcp.insanelib.network.message.EntityModNBTDataSync;
+import insane96mcp.insanelib.network.message.JsonConfigSyncMessage;
+import insane96mcp.insanelib.network.message.MessageCreeperDataSync;
 import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.simple.SimpleChannel;
 
 public class NetworkHandler {
-    private static final String PROTOCOL_VERSION = Integer.toString(1);
-    public static final SimpleChannel CHANNEL = NetworkRegistry.ChannelBuilder
-            .named(new ResourceLocation(InsaneLib.MOD_ID, "simple_channel"))
-            .clientAcceptedVersions(s -> true)
-            .serverAcceptedVersions(s -> true)
-            .networkProtocolVersion(() -> PROTOCOL_VERSION)
-            .simpleChannel();
+    private static final String PROTOCOL_VERSION = Integer.toString(2);
+    public static final SimpleChannel CHANNEL = NetworkRegistry.newSimpleChannel(
+            InsaneLib.location("network_channel"),
+            () -> PROTOCOL_VERSION,
+            PROTOCOL_VERSION::equals,
+            v -> PROTOCOL_VERSION.equals(v) || NetworkRegistry.ABSENT.equals(v) || NetworkRegistry.ACCEPTVANILLA.equals(v)
+    );
+
+    static int index = 0;
 
     public static void init() {
-        CHANNEL.registerMessage(1, MessageCreeperDataSync.class, MessageCreeperDataSync::encode, MessageCreeperDataSync::decode, MessageCreeperDataSync::handle);
-        CHANNEL.registerMessage(2, JsonConfigSyncMessage.class, JsonConfigSyncMessage::encode, JsonConfigSyncMessage::decode, JsonConfigSyncMessage::handle);
+        CHANNEL.registerMessage(++index, MessageCreeperDataSync.class, MessageCreeperDataSync::encode, MessageCreeperDataSync::decode, MessageCreeperDataSync::handle);
+        CHANNEL.registerMessage(++index, JsonConfigSyncMessage.class, JsonConfigSyncMessage::encode, JsonConfigSyncMessage::decode, JsonConfigSyncMessage::handle);
+        CHANNEL.registerMessage(++index, EntityModNBTDataSync.class, EntityModNBTDataSync::encode, EntityModNBTDataSync::decode, EntityModNBTDataSync::handle);
     }
 }