+++ src/main/java/insane96mcp/insanelib/setup/ILStrings.java
@@ -6,9 +6,9 @@ import insane96mcp.insanelib.InsaneLib;
 /// Deprecation: Use {@link insane96mcp.insanelib.module.base.TagsFeature}. This no longer works, it's here to prevent crashes
 public class ILStrings {
 	public static class Tags {
-		public static final String EXPLOSION_CAUSES_FIRE = InsaneLib.RESOURCE_PREFIX + "explosion_causes_fire";
-		public static final String SPAWNED_FROM_SPAWNER = InsaneLib.RESOURCE_PREFIX + "spawned_from_spawner";
-		public static final String SPAWNED_FROM_STRUCTURE = InsaneLib.RESOURCE_PREFIX + "spawned_from_structure";
-		public static final String EXPERIENCE_MULTIPLIER = InsaneLib.RESOURCE_PREFIX + "xp_multiplier";
+		public static final String EXPLOSION_CAUSES_FIRE = InsaneLib.MOD_ID + ":explosion_causes_fire";
+		public static final String SPAWNED_FROM_SPAWNER = InsaneLib.MOD_ID + ":spawned_from_spawner";
+		public static final String SPAWNED_FROM_STRUCTURE = InsaneLib.MOD_ID + ":spawned_from_structure";
+		public static final String EXPERIENCE_MULTIPLIER = InsaneLib.MOD_ID + ":xp_multiplier";
 	}
 }